Refactor analytic
